requirements.txt
chardet
PyYAML
